requirements.pip.txt
pandas==1.3.4
psycopg2==2.9.3
python-dotenv==0.20.0
python_dateutil==2.8.2
requests==2.22.0
